requirements.txt
click==8.1.3
flake8==6.0.0
mccabe==0.7.0
numpy==1.23.5
pandas==1.5.2
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
